backend/requirements/requirements.txt
Django>=3.0,<3.1
wagtail>=2.8,<2.9
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-filter==2.2.0
djangorestframework-jwt==1.11.0
ipstack==0.1.4
django-redis==4.0.0
uWSGI==2.0.18
celery==4.2.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
#psycopg2==2.8.4
